Adding a private member for plugins to attach to RDR's that will be passed back to the plugin when an operation is performed on an RDR so the plugin knows which RDR the operation is refering to.
The previous interface would only let the plugin know the resource, so that
you would get a situation like:

"Plugin, for resource XXXX (via the oid), tell me the sensor reading."

But a given resource can be associated with multiple entities, and each
entity can have multiple sensors, so the plugin would not have enough
information to service the request.

Now when the plug-in allocates a new RDR for a specific sensor associated
with a specific entity on a specific resource, the plug-in will pass a
private data structure along with the RDR structure.  The infrastructure
will forward that private data to the plug-in when the RDR specific plug-in
calls are made.