'list_reources_threaded' is broken into two parts.
First 'NUM_PTHREAD' number of threads begin execution using 'discovery_domain()'.  'discovery_domain()' then performs resource discovery 'NUM_RES_DIS' number of times.  For each loop through resource discovery the present resources discovered are compared to the previous loop's discovered resources.  Inconsistent results are flagged and counted for later output.

After all 'NUM_PTHREAD' threads have completed, by pressing a key a final thread begins execution using 'pthread_event_get()'.  This thread then loops forever or until a second key is pressed.  Each loop of this thread discovers any session events, formats and prints the results.  As with the previous threads the present loop's data is compared to the previous loop's data and any inconsistencies are logged.  This assumes the use of the dummy plug-in that mechanically returns the same session events everytime queried.

On completion of 'list_resources_threaded' any data inconsistencies that have been identified will be displayed.

Additionally if '_TEST_THREADS_' was defined in 'openhpi.h' when the system was built, a count will be maintain and displayed of the number of calls made to 'data_access_lock()' that will block.  This was done to give some indication of the frequency of attempts to access shared data.